feature -- refacto : All-in-One form within multi-pages app: - refacto: all MATCH callbacks defined in AIO scope, all pages callbacks defined in page scope, - feature - refacto: upload-related callbacks were rewritten for AI page, without upload input, TODO: clean it, - feature: title and subtitles added to FormAIO class, TODO: clean it and add more prop as inputs? - centering headers
